+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.1.0] - 2025-01-13
+
+Added JtsPolygonFactory with method to create a bounding box polygon.
+Added JtsPolygonFactoryRijksdriehoek and JtsPolygonFactoryWgs84 as convenient factories for the rijksdriehoek and wgs84 projections.
+
 ## [4.0.0] - 2024-12-12
 
 Breaking change in methods GeodeticCalculatorFactory and BearingCalculator